+++ src/providers/index.ts
@@ -59,6 +59,7 @@ export * from "./utils";
 export { GiteeProvider } from "./gitee";
 export { HuggingFaceProvider } from "./huggingface";
 export { ModelScopeProvider } from "./modelscope";
+export { A4FProvider } from "./a4f";
 
 // 导出模型列表相关函数
 export { getAvailableModels, getAvailableModelsFiltered } from "./registry";